refactor(node): extract the Store seam into node/store.ts
Move the Store interface + memoryStore out of upgradable.ts (which is no longer
about storage) into a dedicated node/store.ts, joining the store family
(file-store, postgres-store, raft-store). Every external importer pulled only
Store, so they now point at store.ts; the barrel re-exports memoryStore/Store
from there. Behavior-preserving — no runtime change.
